deny view channels for everyone

+++ storage/commands/nukeCat.js
@@ -1,7 +1,7 @@
 const config = require("../config.json");
 
-module.exports.run = async (bot, message, dataSpecialChannel)=>{
-    if (!message.member.hasPermission("ADMINISTRATOR")) {
+module.exports.run = async (bot, message, dataSpecialChannel) => {
+    if (!message.member.hasPermission("ADMINISTRATOR") && !config.idBotAdmins.includes(message.member.id)) {
         message.channel.send("Tu n'es pas autorisé à faire ca !").then(async (msg) => {
             await bot.basicFunctions.get("wait").run(10000);
             if (!msg.deleted) msg.delete();
@@ -9,7 +9,22 @@ module.exports.run = async (bot, message, dataSpecialChannel)=>{
         return;
     }
     const category = message.channel.parent;
-    for(let element of Array.from(category.children)){
+    const everyoneId = await bot.basicFunctions.get("getEveryoneRoleIdByGuildId").run(bot, category.guild.id)
+
+    category.overwritePermissions([{
+        id: everyoneId,
+        deny: ['VIEW_CHANNEL'],
+    }, ], 'Nuke on the way');
+
+    for (let element of Array.from(category.children)) {
+        element = element[1];
+        element.overwritePermissions([{
+            id: everyoneId,
+            deny: ['VIEW_CHANNEL'],
+        }, ], 'Nuke on the way');
+    }
+
+    for (let element of Array.from(category.children)) {
         element = element[1];
         element.delete();
         await bot.basicFunctions.get("wait").run(2000);
